task-1/requirements.txt
httpx
bs4
lxml
pytest
celery[redis]
tqdm
fake-useragent
python-dotenv
pandas
flower
